Fixed major bug in TCP framer. Added enums fro message switch. Fixed bug when scrolling and some other events caused server to exit. Found bug - holding and moving mouse fails on low network speed and server hangs. Maybe I should consider using another thread for input

+++ SRD_Server/SRD_Server/ScreenGrabber.hpp
@@ -29,8 +29,8 @@ private :
     CGDirectDisplayID display;
     
     // Should be set via settings
-    size_t output_width = 1920;
-    size_t output_height = 1080;
+    u_long output_width = 1920;
+    u_long output_height = 1080;
     
     uint32_t pixel_format = 'BGRA';
